# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop-api Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "local server"
        }
    ],
    "paths": {
        "/api/cargo-colaboradores/{id}/atualizar-nota": {
            "put": {
                "tags": [
                    "CargoColaborador"
                ],
                "summary": "Atualizar a nota de desempenho de um CargoColaborador",
                "description": "Atualizar a nota de desempenho de um CargoColaborador",
                "operationId": "8f7a6c620c85ba77d303cdfa702aad4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do CargoColaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Corpo da solicitação",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nota_desempenho": {
                                        "description": "Nova nota de desempenho",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota de desempenho atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nota de desempenho atualizada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CargoColaborador não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CargoColaborador não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cargos": {
            "get": {
                "tags": [
                    "Cargos"
                ],
                "summary": "Listar todos os Cargos",
                "description": "Retorna uma lista de todos os Cargos",
                "operationId": "0c650a8167e56e2c6003c4f8272cfa59",
                "responses": {
                    "200": {
                        "description": "Lista de Cargos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cargo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cargos"
                ],
                "summary": "Criar um novo Cargo",
                "description": "Cria um novo Cargo",
                "operationId": "2240790fd53313f002b316739cf1d993",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cargo"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cargo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cargo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cargos/{id}": {
            "get": {
                "tags": [
                    "Cargos"
                ],
                "summary": "Obter detalhes de um Cargo",
                "description": "Retorna detalhes de um Cargo",
                "operationId": "4ae69bb8333baf3faaf454836b0697f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Cargo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do Cargo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cargo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cargo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cargos"
                ],
                "summary": "Atualizar um Cargo",
                "description": "Atualiza um Cargo existente",
                "operationId": "86583ab463268bf39fb7b6c44cb6fdb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Cargo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cargo"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cargo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cargo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cargo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cargos"
                ],
                "summary": "Excluir um Cargo",
                "description": "Exclui um Cargo existente",
                "operationId": "71898049cd50045df81e0ad82d36fadb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Cargo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo excluído com sucesso"
                    },
                    "404": {
                        "description": "Cargo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cargo": {
                "title": "Cargo",
                "description": "Modelo de dados para um Cargo",
                "properties": {
                    "id": {
                        "description": "ID do Cargo",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nome": {
                        "description": "Nome do Cargo",
                        "type": "string"
                    },
                    "descricao": {
                        "description": "Descrição do Cargo",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}